Core separation Phase 2 + Linux/CLI frontend with working REPL
Move socket/resolver/mman to xyzzy-core, split symtable into
core and editor parts, and add xyzzy-cli frontend that links
only xyzzy-core without Win32 GUI dependencies.

Key changes:
- Guard Win32 API calls with #ifdef _WIN32 across 50+ core files
- Add platform.h: Linux compatibility layer (Win32 types via mmap)
- Split gen-syms symtable output into symtable-core.cc/symtable-ed.cc
- Add CLI frontend (cli-main.cc, cli-stubs.cc, cli-conf.cc)
- CMakeLists.txt: Linux/GCC build support, cross-platform xyzzy-cli
- Fix LP64 UB in data.h bit operations (1 -> (u_long)1, int -> u_long)
- Fix EBO overlap in ldata allocator (memset zero-init in do_alloc)

Linux REPL confirmed working: arithmetic, cons, format, strings.
Windows build (xyzzy, xyzzy-batch, xyzzy-cli) verified clean.

+++ src/core/data.h
@@ -85,19 +85,19 @@ struct ldata_free_rep
 inline void
 bitset (u_long *p, int i)
 {
-  p[i / (sizeof *p * CHAR_BIT)] |= 1 << i % (sizeof *p * CHAR_BIT);
+  p[i / (sizeof *p * CHAR_BIT)] |= (u_long)1 << i % (sizeof *p * CHAR_BIT);
 }
 
-inline int
+inline u_long
 bitisset (const u_long *p, int i)
 {
-  return p[i / (sizeof *p * CHAR_BIT)] & (1 << i % (sizeof *p * CHAR_BIT));
+  return p[i / (sizeof *p * CHAR_BIT)] & ((u_long)1 << i % (sizeof *p * CHAR_BIT));
 }
 
 inline void
 bitclr (u_long *p, int i)
 {
-  p[i / (sizeof *p * CHAR_BIT)] &= ~(1 << i % (sizeof *p * CHAR_BIT));
+  p[i / (sizeof *p * CHAR_BIT)] &= ~((u_long)1 << i % (sizeof *p * CHAR_BIT));
 }
 
 int find_zero_bit (u_long *, int);
